requirements.txt
crewai 
crewai-tools
langchain
mysql-connector-python
python-dotenv
